Refactors EnrollmentUIController to ActionButtonComponents - now includes the ApplyButton event for not loggedin users
