+++ frontend-nx/apps/edu-hub/components/CourseContent/ActionButtons/ApplicationModal/FormContent.tsx
